requirements.txt
torch>=1.9.0
Pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.60.0
flask>=2.0.0
gunicorn>=21.0.0
matplotlib>=3.5.0
scikit-image>=0.19.0

